correct decoding of UTF-8 in JSON manifests
Inexplicably, I had commented that we definitely did not want to decode
the octets in JSON.  This statement makes zero sense and I honestly have
no idea what I thought I was saying.

The UTF-8 read from disk was being passed to JSON.pm as if it was already
decoded, meaning that UTF-8 on disk would become mojibake when read.  On
the other hand, \u.... sequences in the JSON would properly become decoded
text.  I wrote a test to demonstrate this bug, then fixed it.

I assume that this was some kind of cognitive overload from when I was
fixing the bigger encoding problems in v3, but I do not care to investigate.